+++ packages/utils/src/fs.test.ts
@@ -1,6 +1,16 @@
 import { promises as fs } from 'fs';
 import { join } from 'path';
-import { readSnapJsonFile, isDirectory, isFile, readJsonFile } from './fs';
+import * as path from 'path';
+import {
+  readSnapJsonFile,
+  isDirectory,
+  isFile,
+  readJsonFile,
+  getOutfilePath,
+  validateOutfileName,
+  validateFilePath,
+  validateDirPath,
+} from './fs';
 import { NpmSnapFileNames } from './types';
 import { DEFAULT_SNAP_BUNDLE, getSnapManifest } from './__test__';
 import { SnapManifest } from './json-schemas';
@@ -149,5 +159,92 @@ describe('readSnapJsonFile', () => {
     await expect(
       readSnapJsonFile(BASE_PATH, NpmSnapFileNames.Manifest),
     ).rejects.toThrow('foo');
+  });
+});
+
+describe('getOutfilePath', () => {
+  it('gets the complete out file path', () => {
+    const path1 = path.normalize('src/outDir');
+    const path2 = path.normalize('../src/outDir/');
+    const path3 = path.normalize('../src/lol/outDir/');
+    const path4 = path.normalize('src/outDir');
+    const path5 = path.normalize('src/outDir/');
+    const path6 = path.normalize('src/bundle.js');
+
+    expect(getOutfilePath('./src', 'outDir')).toStrictEqual(path1);
+    expect(getOutfilePath('../src', '///outDir////')).toStrictEqual(path2);
+    expect(getOutfilePath('../src', '/lol//outDir////')).toStrictEqual(path3);
+    expect(getOutfilePath('src', 'outDir')).toStrictEqual(path4);
+    expect(getOutfilePath('src/', './outDir/')).toStrictEqual(path5);
+    expect(getOutfilePath('src/', '')).toStrictEqual(path6);
+    expect(getOutfilePath('', '')).toStrictEqual('bundle.js');
+  });
+});
+
+describe('validateOutfileName', () => {
+  it('ensures outfile name is just a js file name', () => {
+    expect(() => {
+      validateOutfileName('file.ts');
+    }).toThrow('Invalid outfile name: file.ts');
+
+    expect(() => {
+      validateOutfileName('/');
+    }).toThrow('Invalid outfile name: /');
+
+    expect(() => {
+      validateOutfileName('');
+    }).toThrow('Invalid outfile name: ');
+
+    expect(() => {
+      validateOutfileName('./src/file');
+    }).toThrow('Invalid outfile name: ./src/file');
+
+    expect(() => {
+      validateOutfileName('.js');
+    }).toThrow('Invalid outfile name: .js');
+
+    expect(validateOutfileName('file.js')).toStrictEqual(true);
+    expect(validateOutfileName('two.file.js')).toStrictEqual(true);
+  });
+});
+
+describe('validateFilePath', () => {
+  beforeEach(async () => {
+    await resetFileSystem();
+  });
+
+  it('checks whether the given path string resolves to an existing file', async () => {
+    // jest.spyOn(snapUtils, 'isFile').mockResolvedValue(true);
+    const result = await validateFilePath(MANIFEST_PATH);
+    expect(result).toStrictEqual(true);
+  });
+
+  it('checks whether an invalid path string throws an error', async () => {
+    // jest.spyOn(snapUtils, 'isFile').mockResolvedValue(false);
+    await expect(validateFilePath('/foo/bar.js')).rejects.toThrow(
+      "Invalid params: '/foo/bar.js' is not a file or does not exist.",
+    );
+  });
+});
+
+describe('validateDirPath', () => {
+  beforeEach(async () => {
+    await resetFileSystem();
+  });
+
+  it('checks whether the given path string resolves to an existing directory', async () => {
+    expect(await validateDirPath(BASE_PATH, false)).toBe(true);
+  });
+
+  it('creates the directory if it does not exist and createDir is true', async () => {
+    expect(await validateDirPath(join(BASE_PATH, 'foo'), true)).toBe(true);
+  });
+
+  it('throws if the directory does not exist', async () => {
+    await expect(
+      validateDirPath(join(BASE_PATH, 'bar'), false),
+    ).rejects.toThrow(
+      "Invalid params: '/snap/bar' is not a directory or could not be created.",
+    );
   });
 });